Update README.md
This is the first draft, in upcoming versions I will be adding how to handle sound voices of different lengths. How to reduce noise. Increasing the accuracy of the model, as the data available for speech detection was quite limited. More information on this program will be updated soon covering how amplitudes were extracted, how sampling rates were set with the information about hop length, etc.